requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.5
django-cors-headers==3.7.0
django-rest-framework-mongoengine==3.4.1
django-simpleui==2021.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
mongoengine==0.23.1
Pillow==8.3.1
pycodestyle==2.7.0
PyJWT==2.1.0
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
toml==0.10.2
zipstream==1.1.4
